+++ CHANGELOG.md
@@ -85,3 +85,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - stripHexPrefix fixed (#2989)
 - BatchRequest error handling fixed for callbacks (#2993)
 - ``reconnected`` event and reconnection timeout option added to WebsocketProvider (#2994)
+- Error handling of the HttpProvider fixed (#2887)